Sprint#3 #User Story89 #Task 93 - Create a link to solve questions in the user profile page, only upon selecting their own name from the leader board
